Add organization integrations AI tools page
